requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
ffmpeg-python==0.2.0
filelock==3.17.0
fsspec==2025.2.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
jinja2==3.1.5
llvmlite==0.44.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
pydantic==2.10.6
pydantic-core==2.27.2
pydub==0.25.1
pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
safetensors==0.5.2
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.38
sqlmodel==0.0.22
sse-starlette==2.2.1
starlette==0.45.3
sympy==1.13.1
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typer==0.15.1
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
